Complete deduplication of methods between BaseGitService and HTTPClient interface
- Created generic HTTPClientInterface protocol in http_client.py
- Extracted common HTTP handling logic into provider-specific HTTP client classes:
  - GitHubHTTPClient (updated with generic methods)
  - GitLabHTTPClient (new implementation)
  - BitBucketHTTPClient (new implementation)
- Removed duplicated methods from BaseGitService:
  - execute_request, handle_http_status_error, handle_http_error
  - _has_token_expired, _make_request (now abstract)
- Updated all service classes to inherit from their respective HTTP clients
- Made _fetch_cursorrules_content and _process_microagents_directory abstract in BaseGitService
- Implemented abstract methods in all service classes
- Fixed mypy errors and added proper type annotations
- Added type ignore comments for abstract class instantiation in utils.py

All services (GitHub, GitLab, Bitbucket) now properly inherit from HttpClient interface
and share common HTTP handling logic while maintaining provider-specific implementations.
