+++ .openhands/TASKS.md
@@ -1,9 +1,17 @@
 # Task List
 
-1. ✅ Add constructor to GitHubHTTPClient with initialization logic
+1. ✅ Explore the current codebase structure to understand BaseGitService and HttpClient
 
-2. ✅ Update GitHubService constructor to pass parameters to GitHubHTTPClient
+2. ✅ Identify duplicated methods between BaseGitService and HttpClient interface
 
-3. ✅ Test the changes to ensure functionality is preserved
+3. ✅ Create a generic HTTPClient interface that can be used by all services
 
-4. 🔄 Commit and push the changes
+4. ✅ Refactor BaseGitService to remove duplicated methods and use HttpClient
+
+5. ✅ Update GitLab service class to inherit from HttpClient interface
+
+6. ✅ Update Bitbucket service class to inherit from HttpClient interface
+
+7. 🔄 Test the changes to ensure all services work correctly
+
+8. ⏳ Commit and push the refactored code
