+++ openhands/integrations/bitbucket/http_client.py
@@ -1,24 +1,23 @@
-import json
-from typing import Any, cast
+import base64
+from typing import Any
 
 import httpx
 from pydantic import SecretStr
 
 from openhands.integrations.service_types import (
     RequestMethod,
-    UnknownException,
     User,
 )
 
 
-class GitHubHTTPClient:
+class BitBucketHTTPClient:
     """
-    HTTP client implementation for GitHub API operations.
-    Implements HTTPClientInterface and provides common functionality for GitHub API interactions.
+    HTTP client implementation for Bitbucket API operations.
+    Implements HTTPClientInterface and provides common functionality for Bitbucket API interactions.
     """
 
     BASE_URL: str
-    GRAPHQL_URL: str
+    GRAPHQL_URL: str | None
     token: SecretStr
     refresh: bool
     external_auth_id: str | None
@@ -30,37 +29,38 @@ class GitHubHTTPClient:
         external_auth_id: str | None = None,
         base_domain: str | None = None,
     ) -> None:
-        """Initialize the GitHub HTTP client with configuration."""
+        """Initialize the Bitbucket HTTP client with configuration."""
         # Set default values
-        self.BASE_URL = 'https://api.github.com'
-        self.GRAPHQL_URL = 'https://api.github.com/graphql'
+        self.BASE_URL = 'https://api.bitbucket.org/2.0'
+        self.GRAPHQL_URL = None  # Bitbucket doesn't have GraphQL API
         self.token = token or SecretStr('')
         self.refresh = False
         self.external_auth_id = external_auth_id
-        self.base_domain = base_domain
+        self.base_domain = base_domain or 'bitbucket.org'
 
         # Handle custom domain configuration
-        if base_domain and base_domain != 'github.com':
-            self.BASE_URL = f'https://{base_domain}/api/v3'
-            self.GRAPHQL_URL = f'https://{base_domain}/api/graphql'
+        if base_domain:
+            self.BASE_URL = f'https://api.{base_domain}/2.0'
 
-    async def _get_headers(self) -> dict:
-        """Retrieve the GH Token from settings store to construct the headers."""
-        if not self.token:
-            latest_token = await self.get_latest_token()
-            if latest_token:
-                self.token = latest_token
+    async def _get_headers(self) -> dict[str, str]:
+        """Get headers for Bitbucket API requests."""
+        token_value = self.token.get_secret_value()
 
-        return {
-            'Authorization': f'Bearer {self.token.get_secret_value() if self.token else ""}',
-            'Accept': 'application/vnd.github.v3+json',
-        }
+        # Check if the token contains a colon, which indicates it's in username:password format
+        if ':' in token_value:
+            auth_str = base64.b64encode(token_value.encode()).decode()
+            return {
+                'Authorization': f'Basic {auth_str}',
+                'Accept': 'application/json',
+            }
+        else:
+            return {
+                'Authorization': f'Bearer {token_value}',
+                'Accept': 'application/json',
+            }
 
-    async def _get_github_headers(self) -> dict:
-        """Backward compatibility method."""
-        return await self._get_headers()
-
-    async def get_latest_token(self) -> SecretStr | None:  # type: ignore[override]
+    async def get_latest_token(self) -> SecretStr | None:
+        """Get latest working token of the user."""
         return self.token
 
     async def _make_request(
@@ -68,21 +68,24 @@ class GitHubHTTPClient:
         url: str,
         params: dict | None = None,
         method: RequestMethod = RequestMethod.GET,
-    ) -> tuple[Any, dict]:  # type: ignore[override]
+    ) -> tuple[Any, dict]:
+        """Make a request to the Bitbucket API.
+
+        Args:
+            url: The URL to request
+            params: Optional parameters for the request
+            method: The HTTP method to use
+
+        Returns:
+            A tuple of (response_data, response_headers)
+
+        """
         try:
             async with httpx.AsyncClient() as client:
                 headers = await self._get_headers()
-
-                # Make initial request
                 response = await self._execute_request(
-                    client=client,
-                    url=url,
-                    headers=headers,
-                    params=params,
-                    method=method,
+                    client, url, headers, params, method
                 )
-
-                # Handle token refresh if needed
                 if self.refresh and self._has_token_expired(response.status_code):
                     await self.get_latest_token()
                     headers = await self._get_headers()
@@ -93,14 +96,8 @@ class GitHubHTTPClient:
                         params=params,
                         method=method,
                     )
-
                 response.raise_for_status()
-                response_headers: dict = {}
-                if 'Link' in response.headers:
-                    response_headers['Link'] = response.headers['Link']
-
-                return response.json(), response_headers
-
+                return response.json(), dict(response.headers)
         except httpx.HTTPStatusError as e:
             raise self._handle_http_status_error(e)
         except httpx.HTTPError as e:
@@ -109,46 +106,29 @@ class GitHubHTTPClient:
     async def execute_graphql_query(
         self, query: str, variables: dict[str, Any]
     ) -> dict[str, Any]:
-        try:
-            async with httpx.AsyncClient() as client:
-                headers = await self._get_headers()
-
-                response = await client.post(
-                    self.GRAPHQL_URL,
-                    headers=headers,
-                    json={'query': query, 'variables': variables},
-                )
-                response.raise_for_status()
-
-                result = response.json()
-                if 'errors' in result:
-                    raise UnknownException(
-                        f'GraphQL query error: {json.dumps(result["errors"])}'
-                    )
-
-                return dict(result)
-
-        except httpx.HTTPStatusError as e:
-            raise self._handle_http_status_error(e)
-        except httpx.HTTPError as e:
-            raise self._handle_http_error(e)
+        """Execute a GraphQL query against the API."""
+        # Bitbucket doesn't support GraphQL
+        raise NotImplementedError("Bitbucket doesn't support GraphQL")
 
     async def verify_access(self) -> bool:
+        """Verify that the client has access to the API."""
         url = f'{self.BASE_URL}'
         await self._make_request(url)
         return True
 
-    async def get_user(self):
+    async def get_user(self) -> User:
+        """Get the authenticated user's information."""
         url = f'{self.BASE_URL}/user'
-        response, _ = await self._make_request(url)
+        data, _ = await self._make_request(url)
+
+        account_id = data.get('account_id', '')
 
         return User(
-            id=str(response.get('id', '')),
-            login=cast(str, response.get('login') or ''),
-            avatar_url=cast(str, response.get('avatar_url') or ''),
-            company=response.get('company'),
-            name=response.get('name'),
-            email=response.get('email'),
+            id=account_id,
+            login=data.get('username', ''),
+            avatar_url=data.get('links', {}).get('avatar', {}).get('href', ''),
+            name=data.get('display_name'),
+            email=None,  # Bitbucket API doesn't return email in this endpoint
         )
 
     # Helper methods needed by the HTTP client
@@ -174,15 +154,15 @@ class GitHubHTTPClient:
         if e.response.status_code == 401:
             from openhands.integrations.service_types import AuthenticationError
 
-            return AuthenticationError('Invalid GitHub token')
+            return AuthenticationError('Invalid Bitbucket token')
         elif e.response.status_code == 404:
             from openhands.integrations.service_types import ResourceNotFoundError
 
-            return ResourceNotFoundError(f'Resource not found on GitHub API: {e}')
+            return ResourceNotFoundError(f'Resource not found on Bitbucket API: {e}')
         elif e.response.status_code == 429:
             from openhands.integrations.service_types import RateLimitError
 
-            return RateLimitError('GitHub API rate limit exceeded')
+            return RateLimitError('Bitbucket API rate limit exceeded')
 
         from openhands.integrations.service_types import UnknownException
 
@@ -193,7 +173,3 @@ class GitHubHTTPClient:
         from openhands.integrations.service_types import UnknownException
 
         return UnknownException(f'HTTP error {type(e).__name__} : {e}')
-
-
-# Backward compatibility alias
-GitHubMixinBase = GitHubHTTPClient
